+++ src/main/java/com/marowak/repository/dictonary/SliceTypeRepository.java
@@ -3,5 +3,5 @@ package com.marowak.repository.dictonary;
 import com.marowak.entity.dictonary.SliceType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SliceTypeRepository extends JpaRepository<SliceType, Long> {
+public interface SliceTypeRepository extends JpaRepository<SliceType, Integer> {
 }
